producer/requirements.txt
pandas
requests
python-dotenv
sqlalchemy
psycopg2
cassandra-driver